Proper registration mechanism
Added the whole mechanism. In parallell, I coded the 'server' in a
separate phone (!!) app, I'll publish later after clean up.
Now the registering mechanism allow the app to get a writable view on
the master table, limited to records of the particular user. Changes
uploaded there will be reflected in the base table, and in the public
view that will be used to feed the global vizualizer.